tidy the encryption/decryption with auth renewal awareness
